opt+tcm.base/mv: Move HasOptions out of TCM.Base into Interaction.Options
