+++ src/full/Agda/Interaction/Options.hs
@@ -1,5 +1,7 @@
 module Agda.Interaction.Options
     ( module Opts.Base
+    , module Opts.HasOptions
     ) where
 
 import Agda.Interaction.Options.Base as Opts.Base
+import Agda.Interaction.Options.HasOptions as Opts.HasOptions
